split route
